requirements_complete.txt
# COMPLETE REQUIREMENTS FILE FOR KEMPIAN TALENT PLATFORM
# This file contains ALL dependencies required for deployment

# CORE FLASK FRAMEWORK
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.5
Flask-Migrate>=4.0.5
Flask-CORS>=4.0.0
gunicorn>=21.2.0

# DATABASE
mysqlclient>=2.2.0
PyMySQL>=1.1.0
SQLAlchemy>=2.0.0

# AWS SERVICES
boto3>=1.28.0
botocore>=1.31.0

# AUTHENTICATION
PyJWT>=2.8.0
cryptography>=41.0.0

# PAYMENT
stripe>=6.0.0

# AI/ML CORE
torch>=2.0.0
transformers>=4.33.0
sentence-transformers>=2.2.2
nltk>=3.8.1
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
lightgbm>=4.0.0

# DATA PROCESSING
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# API CLIENTS
requests>=2.31.0
httpx>=0.24.0

# FASTAPI
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# DOCUMENT PROCESSING
PyPDF2>=3.0.0
python-docx>=0.8.11

# UTILITIES
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0

# DEVELOPMENT
pytest>=7.4.0
black>=23.7.0

# DATE/TIME
python-dateutil>=2.8.0
pytz>=2023.3
